+++ flutter/test/mocks.dart
@@ -104,6 +104,7 @@ class MockPlatformChecker with NoSuchMethodProvider implements PlatformChecker {
     this.isRelease = false,
     this.isWebValue = false,
     this.hasNativeIntegration = false,
+    this.isRoot = true,
     Platform? mockPlatform,
   }) : _mockPlatform = mockPlatform ?? MockPlatform('');
 
@@ -111,6 +112,7 @@ class MockPlatformChecker with NoSuchMethodProvider implements PlatformChecker {
   final bool isProfile;
   final bool isRelease;
   final bool isWebValue;
+  final bool isRoot;
   final Platform _mockPlatform;
 
   @override
@@ -124,6 +126,9 @@ class MockPlatformChecker with NoSuchMethodProvider implements PlatformChecker {
 
   @override
   bool isReleaseMode() => isRelease;
+
+  @override
+  bool get isRootZone => isRoot;
 
   @override
   bool get isWeb => isWebValue;